Extract Qt binding compatibility helpers
